+++ assemblies/xipki-qa/src/main/unfiltered/qa/rest.sh
@@ -16,7 +16,7 @@ CA_SHA1FP=`openssl sha1 ${DIR}/../../${CACERT} | cut -d '=' -f 2 | cut -d ' ' -f
 OPTS="--insecure --user user1:password1"
 
 # Use TLS client certificate to authorize
-#OPTS="--insecure --cert ${DIR}/../keycerts/tlskeys/client/tls-client-cert.pem --key ${DIR}/../keycerts/tlskeys/client/tls-client-key.pem"
+#OPTS="--insecure --cert ${DIR}/../keycerts/rest-client-cert.pem --key ${DIR}/../keycerts/rest-client-key.pem"
 
 filename=tls-`date +%s`